Audit: Remove Creature API tests from test/mutate/ (#1769)
Remove AddConnectionOptimisation.ts - 6 of 7 tests tested Creature's
getConnectionSet(), hasConnection(), and getAvailableConnections()
methods rather than AddConnection mutation behaviour. These are already
covered by test/creature/CreatureTopology.ts and
test/creature/SelectiveCacheInvalidation.ts. The remaining mutation
test was a near-duplicate of AvailableConnectionsCache.ts.

Move the unique "excludes constant neurons" test to
AvailableConnectionsCache.ts where it logically belongs.
